MOD: Refactoring

+++ cegui/include/CEGUI/RendererModules/OpenGL/RenderTarget.h
@@ -70,13 +70,13 @@ protected:
     //! holds defined area for the RenderTarget
     Rectf d_area;
     //! tangent of the y FOV half-angle; used to calculate viewing distance.
-    static const double d_yfov_tan;
+    static const float d_yfov_tan;
     //! Projection view matrix cache
     mutable glm::mat4 d_matrix;
     //! true if saved matrix is up to date
     mutable bool d_matrixValid;
     //! tracks viewing distance (this is set up at the same time as d_matrix)
-    mutable double d_viewDistance;
+    mutable float d_viewDistance;
 };
 
 }
